speed is dyn now separed from decay, changed speed scale

+++ videoCutOnset.py
@@ -7,14 +7,13 @@ if len(sys.argv)>3:
     mindur = float(sys.argv[3])
 
 name = os.path.basename(file)
-extension = os.path.splitext(name)[1]
+extension = osq.path.splitext(name)[1]
 
-os.system("ffmpeg -i "+sys.argv[1]+" -map 0:a "+name+".audio.wav -map 0:v "+name+".onlyvideo.avi")
-os.system("rm "+name+".onlyvideo.avi")
-os.system("aubiocut "+name+".audio.wav > "+name+".onsets")
-os.system("aubiocut -S  "+name+".audio.wav > "+name+".onsets")
-
-os.system("rm "+name+".audio.wav")
+if not os.path.isfile(name+".onsets"):
+    os.system("ffmpeg -i "+sys.argv[1]+" -map 0:a "+name+".audio.wav -map 0:v "+name+".onlyvideo.avi")
+    os.system("rm "+name+".onlyvideo.avi")
+    os.system("aubiocut "+name+".audio.wav > "+name+".onsets")
+    os.system("rm "+name+".audio.wav")
 
 f = open(name+".onsets","r"); o = [float(line) for line in f.readlines()];
